requirements.txt
Django==1.6.5
django-extensions==1.5.5
django-cors-headers==0.12
django-filter==0.8
djangorestframework==2.3.14
django-rest-swagger==0.1.14
saga-python==0.14
radical.utils==0.7.11
